# Add indexes to improve patient_summaries performance