--- a/db/migrate/20200430141834_add_scheduled_date_index.rb
+++ b/db/migrate/20200430141834_add_scheduled_date_index.rb
@@ -1,0 +1,5 @@
+class AddScheduledDateIndex < ActiveRecord::Migration[5.1]
+  def change
+    add_index :appointments, [:patient_id, :scheduled_date], order: {patient_id: :asc, scheduled_date: :desc}
+  end
+end