--- a/db/migrate/20200430144041_add_blood_pressure_index.rb
+++ b/db/migrate/20200430144041_add_blood_pressure_index.rb
@@ -1,0 +1,5 @@
+class AddBloodPressureIndex < ActiveRecord::Migration[5.1]
+  def change
+    add_index :blood_pressures, [:patient_id, :recorded_at], order: {patient_id: :asc, recorded_at: :desc}
+  end
+end