(cherry picked from https://github.com/simpledotorg/simple-server/pull/907 to keep schema changes independent)

These two indexes cut down significantly on time spent in two `order
by` clauses we are doing for the patient_summaries view.

These indexes are compound and need to sort the second date field descending
to match how we commonly retrieve appointments and blood_pressure.

On the main overdue list page this cut the overall page load time by over half:

**Before**
```
Completed 200 OK in 8449ms (Views: 309.2ms | ActiveRecord: 8119.7ms)
```

**After**
```
Completed 200 OK in 3080ms (Views: 192.8ms | ActiveRecord: 2869.1ms)
```

I used the [Postgres Query Analyzer](http://tatiyants.com/pev/#/plans) to find the worst parts of the query plan, and then to introduce indices to make sure it brought those times down. It was super handy.  Here is are the query plans from that tool, with before and after. The key part to notice is that the two costliest parts of the query go from **SEQ SCAN** to **INDEX SCAN**, and it also drops out the sort operation (as the index is already sorted how we want).

**Before**

![Screen Shot 2020-04-30 at 2 39 53 PM](https://user-images.githubusercontent.com/69/80752069-8799a900-8af0-11ea-9d69-11d2f037c850.png)

**After**

![Screen Shot 2020-04-30 at 2 41 13 PM](https://user-images.githubusercontent.com/69/80752175-ad26b280-8af0-11ea-9f26-c84d414d2591.png)

**Story card:** https://www.pivotaltracker.com/n/projects/2184102/stories/172537522/

## Because

We want the dashboards to feel fast

## This addresses

The query time of the patient_summaries view.